requirements.txt
selenium
requests
pillow
colorama
tqdm
numpy
tensorflow-lite
